requirements.txt
streamlit==1.55.0
requests==2.31.0
python-dotenv==1.0.0
# Optional: for fetching real market data
# yfinance==0.2.34
